all good

+++ QuickTCGA/Logic/QuickTCGASegmenter.h
@@ -11,26 +11,42 @@
 #include "opencv2/core/core.hpp"
 #include "opencv2/highgui/highgui.hpp"
 #include "opencv2/imgproc/imgproc.hpp"
-
 #include "TCGAUtilities.h"
-#include "./NucleusSeg_Yi/ProcessTileUtils.h"
+//#include "NucleusSeg_Yi/ProcessTileUtils.h"
+#include "NucleusSeg_Yi/utilityTileAnalysis.h"
 
 class QuickTCGASegmenter {
 public:
     QuickTCGASegmenter();
+
     ~QuickTCGASegmenter();
 
-    void SetSourceImage(const cv::Mat& imSrc);
-    void SetLabImage(const cv::Mat& imLab);
-    void SetROIImage(const cv::Mat& imROI);
-    void SetPreSegmentation(const cv::Mat& imSeg);
+    void SetSourceImage(const cv::Mat &imSrc);
+
+    void SetLabImage(const cv::Mat &imLab);
+
+    void SetROIImage(const cv::Mat &imROI);
+
+    void SetPreSegmentation(const cv::Mat &imSeg);
+
     void DoSegmentation();
+
     void DoTemplateMatching();
-    void DoNucleiSegmentationYi(float otsuRatio, double curvatureWeight, float sizeThld, float sizeUpperThld, double mpp, double kernelSize);
-    void DoNucleiSegmentationYiwo(float otsuRatio, double curvatureWeight, float sizeThld, float sizeUpperThld, double mpp, double kernelSize);
-    void GetSegmentation(cv::Mat& imSeg);
+
+    void
+    DoNucleiSegmentationYi(float otsuRatio, double curvatureWeight, float sizeThld, float sizeUpperThld, double mpp,
+                           float kernelSize);
+
+    void
+    DoNucleiSegmentationYiwo(float otsuRatio, double curvatureWeight, float sizeThld, float sizeUpperThld, double mpp,
+                             float kernelSize);
+
+    void GetSegmentation(cv::Mat &imSeg);
+
     void RefineCurvature();
+
     void RefineShortCut();
+
     void OnTrackColorThreshold(int indColorTh);
 
 private:
@@ -51,8 +67,8 @@ private:
     static const int m_COUNTOURAREA_MIN;
     static const int m_COLOR_SLIDER_MAX;
     static const int m_STRELE_SIZE;
-    std::vector<std::vector<cv::Point> > m_contours;
-    std::vector<cv::Vec4i> m_hierarchy;
+    std::vector <std::vector<cv::Point> > m_contours;
+    std::vector <cv::Vec4i> m_hierarchy;
 };
 
 #endif
